Refactoring

+++ pytorch_lightning/metrics/classification/stat_scores.py
@@ -165,7 +165,7 @@ class StatScores(Metric):
                 zeros_shape = []
             elif reduce == "macro":
                 zeros_shape = (num_classes,)
-            default, reduce_fn = lambda: torch.zeros(zeros_shape, dtype=torch.int), "sum"
+            default, reduce_fn = lambda: torch.zeros(zeros_shape, dtype=torch.long), "sum"
         else:
             default, reduce_fn = lambda: [], None
